Conflicts:
	python/CMakeLists.txt
	python/extractor_dict.py
	releaseDB/base_dependency_database
	scripts/condor_lar.sh
	scripts/dropbox.py
	scripts/make_xml_mcc5.0.sh
	scripts/project.py
	scripts/rootstat.py
	uboone/AnalysisTree/AnalysisTree_module.cc
	uboone/AnalysisTree/anatree.fcl
	uboone/Geometry/gdml/microboone_overburden.gdml
	uboone/Geometry/gdml/microboone_overburden_geo.C
	uboone/Geometry/gdml/microboone_overburden_nowires.gdml
	uboone/MCShowerReco/MCShowerRecoAlg.cxx
	uboone/MCShowerReco/MCShowerRecoEdep.cxx
	uboone/MCShowerReco/ToyOneShowerGen_module.cc
	ups/product_deps
	xml/mcc5.0/prodcosmics_uboone.xml
	xml/mcc5.0/prodgenie_bnb_nue_cosmic_uboone.xml
	xml/mcc5.0/prodgenie_bnb_nue_uboone.xml

+++ uboone/MCShowerReco/MCShowerRecoEdep.h
@@ -74,9 +74,9 @@ namespace sim
     float y;
     float z;
     float energy;
-    float qU;
-    float qV;
-    float qW;
+    double qU;
+    double qV;
+    double qW;
     std::map<unsigned short,sim::MCShowerHit> mchits;
     void Clear() {
       //x=y=z=kINVALID_SHORT;
